// src/models/account.model.ts
export default class AccountModel {
    Id: string = '';
    UserId: string = '';
    UserName: string = '';
    Password: string = '';
    OldPassword: string = '';
    CreatedDate: Date = new Date();
    CreateBy: string = '';
    StatusId: number = 0;
    StatusName: string = '';
    FirstName: string = '';
    LastName: string = '';
    Email: string = '';
    Address: string = '';
    Phone: string = '';
    IdentityNo: string = '';
}